+++ lib/recipes/pl-marketing/recipe.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-PL_MARKETING_RECIPE = Kitchen::BookRecipe.new(book_short_name: :plmarketing) do |doc, resources|
+PL_MARKETING_RECIPE = Kitchen::BookRecipe.new(book_short_name: :plmarketing) do |doc, _resources|
   include Kitchen::Directions
 
   # Set overrides
@@ -14,7 +14,6 @@ PL_MARKETING_RECIPE = Kitchen::BookRecipe.new(book_short_name: :plmarketing) do 
   book.search('cnx-pi').trash
 
   BakeListsWithPara.v1(book: book)
-  BakeImages.v1(book: book, resources: resources)
   BakePreface.v1(book: book, cases: true)
 
   book.pages('$.preface').each do |page|