+++ lib/recipes/web-generic/recipe.rb
@@ -3,7 +3,16 @@
 WEB_RECIPE = Kitchen::BookRecipe.new(book_short_name: :web) do |doc, resources|
   include Kitchen::Directions
 
+  # Setup
   book = doc.book
+  book_pages = book.search_with(
+    Kitchen::PageElementEnumerator, Kitchen::CompositePageElementEnumerator
+  )
+  WebPreBakeSetup.v1(book_pages: book_pages)
 
-  BakeImages.v1(book: book, resources: resources)
+  # Web directions
+  BakeImages.v1(book_pages: book_pages, resources: resources)
+
+  # Restore
+  WebPostBakeRestore.v1(book_pages: book_pages)
 end