+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Create `BakeOrderHeaders` direction & specs
 * Bake `chapter outline` in `neuroscience`
 
 ## [v2.7.0] - 2023-12-01