+++ CHANGELOG.md
@@ -6,10 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Add `tables` in `neuroscience`
+* Added `index` and `appendix` to `neuroscience`
 * Bake `lo` in `data-science`
 * Add `developmental-perspective` note in `neuroscience`
 * Create basic files for `data-science` recipe
-* Added `index` and `appendix` to `neuroscience`
 * Create `BakeOrderHeaders` direction & specs
 * Bake `chapter outline` in `neuroscience`