+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Bake `EOC` in `data-science`
+* Add `tables` in `neuroscience`
+* Added `index` and `appendix` to `neuroscience`
 * Bake `lo` in `data-science`
 * Add `developmental-perspective` note in `neuroscience`
 * Create basic files for `data-science` recipe
@@ -26,10 +28,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Bake `figures` in `lifespan-development`
 * Create basic files for `lifespan-development` recipe
 
+
 ## [v2.6.0] - 2023-11-17
 
 * Add `boxed-feature` note in `neuroscience`
 * Bake EOC sections in  `neuroscience`
+* Add `tables` in `neuroscience`
 * Add `neurscience across species` note in `neuroscience`
 * Bake `learning-objectives` in `neuroscience`
 * Add `BakeRexWrappers` to every recipe-- gives an attribute for rex