+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.8.0] - 2023-12-15
+
 * Create `WebPreBakeSetup` and `WebPostBakeRestore` directions
 * Add `WebPreBakeSetup` and `WebPostBakeRestore` to web recipe
 * Update web recipe to use `book_pages` instead of `book`